requirements.txt
accelerate==0.15.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
diffusers==0.10.2
filelock==3.8.2
fonttools==4.38.0
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==5.1.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.5
opencv-python==4.6.0.66
packaging==22.0
Pillow==9.3.0
psutil==5.9.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
six==1.16.0
tokenizers==0.13.2
torch==1.13.0
tqdm==4.64.1
transformers==4.25.1
typing-extensions==4.4.0
urllib3==1.26.13
zipp==3.11.0
